refactor: 스크롤바 로직 custom hooks로 추상화
